+++ lib/src/parse/stylesheet.dart
@@ -1534,7 +1534,7 @@ relase. For details, see http://bit.ly/moz-document.
 
     resolveOperations() {
       if (operators == null) return;
-      while (!operators.isEmpty) {
+      while (operators.isNotEmpty) {
         resolveOneOperation();
       }
     }